% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/df_rep_of_each_level_a_factor_in_each_level_another_factor.R
\name{df_rep_of_each_level_a_factor_in_each_level_another_factor}
\alias{df_rep_of_each_level_a_factor_in_each_level_another_factor}
\title{df_rep_of_each_level_a_factor_in_each_level_another_factor}
\usage{
df_rep_of_each_level_a_factor_in_each_level_another_factor(
  df,
  statcol,
  otherfaccol,
  optionrep,
  personnal_colors = FALSE,
  colchoose = NULL
)
}
\arguments{
\item{df}{a dataframe (class data.frame) with at least two columns of class character or class factor. These columns should give the level of a specific factor for each sample (row) of the dataframe.}

\item{statcol}{an element of class character that should correspond to the name of a column of class character or factor of the dataframe df.}

\item{otherfaccol}{an element of class character that should correspond to the name of a column of class character or factor of the datarame df. The value of the parameter otherfaccol 
should not be the same as the value of the parameter statcol.}

\item{optionrep}{an element of class character that should correspond either to percentage or brut. In the first case the distribution of the levels of the factor otherfaccol inside the levels of
the factor statcol is represented in percentage otherwise when the value of optionrep is brut representing  the real number of samples.}

\item{personnal_colors}{an element of class logical. if TRUE  the colors to represent the levels of the factor otherfaccol are provided by the user otherwise different colors choosen according to the number of levels of the factor otherfaccol will be used (determined with the package chameleon).}

\item{colchoose}{a vector of class character. It provides the color that will be associated to each level of the factor otherfaccol if the parameter personnal_colors is set to TRUE.}
}
\value{
generate a ggplot2 barplot representation that gives the distribution of the levels of the factor otherfaccol into the level of the factor statcol.
}
\description{
This function allows to make a barplot representation of the proportions of the different level of a factor into another factor.
}
\examples{
require(Seurat)



df_rep_of_each_level_a_factor_in_each_level_another_factor(processPBMC@meta.data,
statcol = "seurat_clusters",
otherfaccol = "Phase",
personnal_colors = FALSE,
optionrep = "percentage")

df_rep_of_each_level_a_factor_in_each_level_another_factor(processPBMC@meta.data,
statcol = "seurat_clusters",
otherfaccol = "Phase",
personnal_colors = TRUE,
colchoose = c("S" = "blue","G1" = "green","G2M" = "red"),
optionrep = "percentage")


}
