% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IndiceRepSignaturePval.R
\name{IndiceRepSignaturePval}
\alias{IndiceRepSignaturePval}
\title{IndiceRepSignaturePval}
\usage{
IndiceRepSignaturePval(
  exprMat,
  biological_signature,
  min_prop_cells_detected,
  number_random_signatures = 100,
  thresholdpval,
  random_seed
)
}
\arguments{
\item{exprMat}{object of class matrix that corresponds to the single-cell expression matrix.}

\item{biological_signature}{a character vector that corresponds to a list of genes of interest.}

\item{min_prop_cells_detected}{Numerical value that corresponds to the minimal proportion of genes that have to be detected to take into account a gene in the matrix.}

\item{number_random_signatures}{integer value. The number of random signatures that will be used to generate the null distribution.}

\item{thresholdpval}{numeric value which corresponds to the p-value threshold applied to determine whether or not a cell is significantly enriched for a given biological signature.}

\item{random_seed}{integer value random seed to ensure the reproducibility of the results}
}
\value{
(named list). This function returns a named list : each element of this list is a numeric vector that gives values for each droplet of a specific metric except the last element, a character #'   vector that gives the status of each cell.
The fist element of list called ratio_interest_sign_vs_random provides for each droplet the ratio between the percentage of detected genes in the interest signature and the mean percentage of detected #' genes in random signatures of the same size than the interest signature.
The second element of the list called detection_rate_signature gives the percentage of detected genes in the interest signature.      
The third element of the list called mean_random_detection_rate_per_cell gives the mean percentage of detected genes in all the random signatures.
The fourth element called p-value gives the numeric value (p-value) that allow to assess if the detection rate of the cell is significantly higher or lower than expected randomly. To distinguish the #'  two cases (lower or higher) when the percentage of detection rate in the interest signature is higher than the median percentage of detection rate, the sign of the value is positive. In the #'  
alternative case (percentage of detected genes in the interest signature is lower than the median percentage of detected genes in the random signatures)  
The fifth element called abs_pvalue is the same thing as the fourth element (p-value) but all the values are positive (original values).
The sixth element called statut_pvalue is a character vector that gives the status of each cell according to the value of the fourth element : POS_SIGNI (detection rate higher than expected randomly), #' NEG_SIGNI (detection rate lower than expected randomly) or NOT_SIGNI (detection rate not significantly different than expected randomly).
}
\description{
this function returns a named list : each element of this list is a numeric vector that gives values for each droplet of a specific metric except the last element, a character vector that #' gives the status of each cell. The fist element of list called ratio_interest_sign_vs_random provides for each droplet the ratio between the percentage of detected genes in the interest signature and #' the mean percentage of detected genes in random signatures of the same size than the interest signature.
The second element of the list called detection_rate_signature gives the percentage of detected genes in the interest signature.      
The third element of the list called mean_random_detection_rate_per_cell gives the mean percentage of detected genes in all the random signatures.
The fourth element called p-value gives the numeric value (p-value) that allow to assess if the detection rate of the cell is significantly higher or lower than expected randomly. To distinguish the #'  two cases (lower or higher) when the percentage of detection rate in the interest signature is higher than the median percentage of detection rate, the sign of the value is positive. In the 
 alternative case (percentage of detected genes in the interest signature is lower than the median percentage of detected genes in the random signatures)  
The fifth element called abs_pvalue is the same thing as the fourth element (p-value) but all the values are positive (original values).
The sixth element called statut_pvalue is a character vector that gives the status of each cell according to the value of the fourth element : POS_SIGNI (detection rate higher than expected randomly), #' NEG_SIGNI (detection rate lower than expected randomly) 
or NOT_SIGNI (detection rate not significantly different than expected randomly).
}
\examples{

require(Seurat)



platelet <- read.table(GenicSignature_example("platelet_signature"),header=FALSE)[[1]]

results_platelets <- IndiceRepSignaturePval(exprMat = processPBMC@assays$SCT@counts,
biological_signature = platelet,
min_prop_cells_detected = 0.02,
number_random_signatures = 1000,
thresholdpval = 0.01,
random_seed = 1)

processPBMC@meta.data$status_platelets <- results_platelets$statut_fdr

processPBMC$platelets_ratio_query_sign_vs_random <- results_platelets$ratio_interest_sign_vs_random


Seurat::DimPlot(processPBMC,group.by= "status_platelets")


Seurat::FeaturePlot(processPBMC,"platelets_ratio_query_sign_vs_random")

}
