% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbmc_example.R
\name{pbmc_example}
\alias{pbmc_example}
\title{pbmc_example}
\usage{
pbmc_example(data = "pbmc")
}
\arguments{
\item{data}{Name of the example dataset.}
}
\description{
Get path to files to run the scSIGTest method on the pbmc dataset.

scSIGTest comes bundled with some example files in its `inst/extdata`
directory. This function allows to load files to run the scSIGTest method for example on the pbmc dataset
}
\examples{

pbmc_example("pbmc")

}
