% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessPBMC.R
\name{ProcessPBMC}
\alias{ProcessPBMC}
\title{ProcessPBMC}
\usage{
ProcessPBMC()
}
\value{
A Seurat object containing the PBMC already basically processed and ready to use as an example to illustrate how some functions of the package work.
}
\description{
Perform a basic preprocessing of the PBMC dataset which is used in this package to provide some examples of use of some of the defined function.
}
\examples{

require(Seurat)



Seurat::DimPlot(processPBMC)


}
